+++ src/app/store/actions/types/auth.actiontypes.ts
@@ -43,17 +43,6 @@ export enum ActionTypes {
   DELETE_ACCOUNT = '[Auth] Delete Account',
   DELETE_ACCOUNT_SUCCESS = '[Auth] Delete Account Success',
   DELETE_ACCOUNT_FAILURE = '[Auth] Delete Account Failure',
-  BLOCK_USER = '[User] Block User',
-  BLOCK_USER_SUCCESS = '[User] Block User Success',
-  BLOCK_USER_FAILURE = '[User] Block User Failure',
-  BLOCK_USER_INITIAL_STATE = '[User] Block User Initial State',
-  UNBLOCK_USER = '[User] Unblock User',
-  UNBLOCK_USER_SUCCESS = '[User] Unblock User Success',
-  UNBLOCK_USER_FAILURE = '[User] Unblock User Failure',
-  UNBLOCK_USER_INITIAL_STATE = '[User] Unblock User Initial State',
-  GET_BLOCKED_USERS = '[User] Get Blocked Users',
-  GET_BLOCKED_USERS_SUCCESS = '[User] Get Blocked Users Success',
-  GET_BLOCKED_USERS_FAILURE = '[User] Get Blocked Users Failure',
   LOGOUT = '[Auth] Logout',
   LOGOUT_SUCCESS = '[Auth] Logout Success',
   LOGOUT_FAILURE = '[Auth] Logout Failure',
