+++ src/app/store/actions/types/user.actiontypes.ts
@@ -12,4 +12,15 @@ export enum ActionTypes {
   REPORT_USER_SUCCESS = '[User] Report User Success',
   REPORT_USER_FAILURE = '[User] Report User Failure',
   REPORT_USER_INITIAL_STATE = '[User] Report User Initial State',
+  BLOCK_USER = '[User] Block User',
+  BLOCK_USER_SUCCESS = '[User] Block User Success',
+  BLOCK_USER_FAILURE = '[User] Block User Failure',
+  BLOCK_USER_INITIAL_STATE = '[User] Block User Initial State',
+  UNBLOCK_USER = '[User] Unblock User',
+  UNBLOCK_USER_SUCCESS = '[User] Unblock User Success',
+  UNBLOCK_USER_FAILURE = '[User] Unblock User Failure',
+  UNBLOCK_USER_INITIAL_STATE = '[User] Unblock User Initial State',
+  GET_BLOCKED_USERS = '[User] Get Blocked Users',
+  GET_BLOCKED_USERS_SUCCESS = '[User] Get Blocked Users Success',
+  GET_BLOCKED_USERS_FAILURE = '[User] Get Blocked Users Failure',
 }
