+++ src/app/store/actions/user.action.ts
@@ -73,3 +73,59 @@ export const reportUserFailureAction = createAction(
 export const reportUserInitialStateAction = createAction(
   ActionTypes.REPORT_USER_INITIAL_STATE
 );
+
+// Block User
+export const blockUserAction = createAction(
+  ActionTypes.BLOCK_USER,
+  props<{ request: { userId: string } }>()
+);
+
+export const blockUserSuccessAction = createAction(
+  ActionTypes.BLOCK_USER_SUCCESS,
+  props<{ payload: User }>()
+);
+
+export const blockUserFailureAction = createAction(
+  ActionTypes.BLOCK_USER_FAILURE,
+  props<{ error: ErrorInterface }>()
+);
+
+export const blockUserInitialStateAction = createAction(
+  ActionTypes.BLOCK_USER_INITIAL_STATE
+);
+
+// Unblock User
+export const unBlockUserAction = createAction(
+  ActionTypes.UNBLOCK_USER,
+  props<{ request: { userId: string } }>()
+);
+
+export const unBlockUserSuccessAction = createAction(
+  ActionTypes.UNBLOCK_USER_SUCCESS,
+  props<{ payload: User }>()
+);
+
+export const unBlockUserFailureAction = createAction(
+  ActionTypes.UNBLOCK_USER_FAILURE,
+  props<{ error: ErrorInterface }>()
+);
+
+export const unBlockUserInitialStateAction = createAction(
+  ActionTypes.UNBLOCK_USER_INITIAL_STATE
+);
+
+// Get Blocked Users
+export const getBlockedUsersAction = createAction(
+  ActionTypes.GET_BLOCKED_USERS,
+  props<{ request: { blockedUsers: string[] } }>()
+);
+
+export const getBlockedUsersSuccessAction = createAction(
+  ActionTypes.GET_BLOCKED_USERS_SUCCESS,
+  props<{ payload: User[] }>()
+);
+
+export const getBlockedUsersFailureAction = createAction(
+  ActionTypes.GET_BLOCKED_USERS_FAILURE,
+  props<{ error: ErrorInterface }>()
+);
