+++ src/app/store/effects/user.effect.ts
@@ -32,6 +32,15 @@ import {
   updateCurrentUserAction,
   updateCurrentUserFailureAction,
   updateCurrentUserSuccessAction,
+  blockUserAction,
+  blockUserFailureAction,
+  blockUserSuccessAction,
+  getBlockedUsersAction,
+  getBlockedUsersFailureAction,
+  getBlockedUsersSuccessAction,
+  unBlockUserAction,
+  unBlockUserFailureAction,
+  unBlockUserSuccessAction,
 } from 'src/app/store/actions/user.action';
 
 @Injectable()
@@ -125,6 +134,64 @@ export class UserEffects {
               return of(reportUserFailureAction({ error }));
             })
           );
+      })
+    )
+  );
+
+  blockUser$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(blockUserAction),
+      withLatestFrom(this.store.pipe(select(currentUserSelector))),
+      switchMap(([{ request }, currentUser]) => {
+        return this.userService.blockUser(currentUser, request.userId).pipe(
+          map((payload: User) => blockUserSuccessAction({ payload })),
+
+          catchError((errorResponse: HttpErrorResponse) => {
+            const error: ErrorInterface = {
+              message: errorResponse.message,
+            };
+            return of(blockUserFailureAction({ error }));
+          })
+        );
+      })
+    )
+  );
+
+  unBlockUser$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(unBlockUserAction),
+      withLatestFrom(this.store.pipe(select(currentUserSelector))),
+      switchMap(([{ request }, currentUser]) => {
+        return this.userService.unBlockUser(currentUser, request.userId).pipe(
+          map((payload: User) => unBlockUserSuccessAction({ payload })),
+
+          catchError((errorResponse: HttpErrorResponse) => {
+            const error: ErrorInterface = {
+              message: errorResponse.message,
+            };
+            return of(unBlockUserFailureAction({ error }));
+          })
+        );
+      })
+    )
+  );
+
+  getBlockedUsers$ = createEffect(() =>
+    this.actions$.pipe(
+      ofType(getBlockedUsersAction),
+      switchMap(({ request }) => {
+        return this.userService.getBlockedUsers(request.blockedUsers).pipe(
+          map((payload: User[]) => {
+            return getBlockedUsersSuccessAction({ payload });
+          }),
+
+          catchError((errorResponse: HttpErrorResponse) => {
+            const error: ErrorInterface = {
+              message: errorResponse.message,
+            };
+            return of(getBlockedUsersFailureAction({ error }));
+          })
+        );
       })
     )
   );
