+++ src/app/store/reducers/auth.reducer.ts
@@ -50,17 +50,7 @@ import {
   deleteAccountAction,
   deleteAccountSuccessAction,
   deleteAccountFailureAction,
-  blockUserAction,
-  blockUserSuccessAction,
-  blockUserFailureAction,
-  blockUserInitialStateAction,
-  getBlockedUsersAction,
-  getBlockedUsersSuccessAction,
-  getBlockedUsersFailureAction,
-  unBlockUserAction,
-  unBlockUserFailureAction,
-  unBlockUserSuccessAction,
-  unBlockUserInitialStateAction,
+
 } from 'src/app/store/actions/auth.action';
 import {
   updatePresenceFailureAction,
@@ -73,6 +63,17 @@ import {
   updateCurrentUserAction,
   updateCurrentUserFailureAction,
   updateCurrentUserSuccessAction,
+  blockUserAction,
+  blockUserSuccessAction,
+  blockUserFailureAction,
+  blockUserInitialStateAction,
+  getBlockedUsersAction,
+  getBlockedUsersSuccessAction,
+  getBlockedUsersFailureAction,
+  unBlockUserAction,
+  unBlockUserFailureAction,
+  unBlockUserSuccessAction,
+  unBlockUserInitialStateAction,
 } from 'src/app/store/actions/user.action';
 import {
   createLanguageAction,
